few days of work

+++ Iris based Unibiometric system/match_biometric.m
@@ -1,3 +1,3 @@
 function score=match_biometric(query,template)
-    score=1-norm(template-query)/(norm(template)+norm(query));
+    score=1-norm(double(template)-double(query))/(norm(double(template))+norm(double(query)));
 end
